rename aes module to symmetric for correctnes
