+++ src/crypto/symmetric/mod.rs
@@ -1,3 +1,3 @@
-pub mod aes;
 pub mod encryption;
 pub mod key;
+pub mod xsalsapoly;
